Moving path normalization methods to command_line_utils.py
