+++ test/test_dynamic_service_helper.py
@@ -737,77 +737,6 @@ class TestProcess:
         assert p.pobjectid.time_observed == "1970-01-01 00:00:02.000"
         assert p.pobjectid.session == "blah"
 
-    @staticmethod
-    @pytest.mark.parametrize(
-        "path, expected_result",
-        [
-            ("blah", "x86"),
-            ("C:\\program files\\blah", "x86"),
-            ("C:\\program files (x86)\\blah", "x86_64"),
-            ("C:\\syswow64\\blah", "x86_64"),
-        ],
-    )
-    def test_determine_arch(path, expected_result):
-        current_objectid = ObjectID(tag="blah", ontology_id="blah", service_name="blah")
-        p = Process(
-            objectid=current_objectid, image=path, start_time="1970-01-01 00:00:00.001"
-        )
-        actual_result = p._determine_arch(path)
-        assert actual_result == expected_result
-
-    @staticmethod
-    @pytest.mark.parametrize(
-        "path, rule, expected_result",
-        [
-            ("blah", {"pattern": "", "replacement": ""}, "blah"),
-            ("blah", {"pattern": "ah", "replacement": "ue"}, "blah"),
-            ("blah", {"pattern": "bl", "replacement": "y"}, "yah"),
-        ],
-    )
-    def test_pattern_substitution(path, rule, expected_result):
-        actual_result = Process._pattern_substitution(path, rule)
-        assert actual_result == expected_result
-
-    @staticmethod
-    @pytest.mark.parametrize(
-        "path, rule, expected_result",
-        [
-            ("blah", {"regex": "", "replacement": ""}, "blah"),
-            ("blah", {"regex": "bl*ah", "replacement": "bl"}, "blah"),
-            ("blah", {"regex": "\\bl*ah", "replacement": "bl"}, "blah"),
-            ("blaah", {"regex": "bl*ah", "replacement": "blue"}, "blue"),
-        ],
-    )
-    def test_regex_substitution(path, rule, expected_result):
-        actual_result = Process._regex_substitution(path, rule)
-        assert actual_result == expected_result
-
-    @staticmethod
-    @pytest.mark.parametrize(
-        "path, arch, expected_result",
-        [
-            ("blah", None, "blah"),
-            ("C:\\Program Files\\Word.exe", None, "?pf86\\word.exe"),
-            ("C:\\Program Files (x86)\\Word.exe", None, "?pf86\\word.exe"),
-            ("C:\\Program Files (x86)\\Word.exe", "x86_64", "?pf86\\word.exe"),
-            ("C:\\Windows\\System32\\Word.exe", None, "?sys32\\word.exe"),
-            ("C:\\Windows\\SysWow64\\Word.exe", None, "?sys32\\word.exe"),
-            ("C:\\Windows\\SysWow64\\Word.exe", "x86", "?win\\syswow64\\word.exe"),
-            ("C:\\Windows\\SysWow64\\Word.exe", "x86_64", "?sys32\\word.exe"),
-            (
-                "C:\\Users\\buddy\\AppData\\Local\\Temp\\Word.exe",
-                None,
-                "?usrtmp\\word.exe",
-            ),
-            ("C:\\Users\\buddy\\Word.exe", None, "?usr\\word.exe"),
-            ("%WINDIR%\\explorer.exe", None, "?win\\explorer.exe"),
-            ("%SAMPLEPATH%\\diagerr.exe", None, "?usrtmp\\diagerr.exe"),
-        ],
-    )
-    def test_normalize_path(path, arch, expected_result):
-        actual_result = Process._normalize_path(path, arch)
-        assert actual_result == expected_result
-
 
 class TestNetworkConnection:
     @staticmethod
